Extract the common meta utils
