+++ tools/pd-heartbeat-bench/main.go
@@ -18,7 +18,6 @@ import (
 	"context"
 	"crypto/tls"
 	"fmt"
-	"io"
 	"math/rand"
 	"net/http"
 	"os"
@@ -46,21 +45,15 @@ import (
 	pdHttp "github.com/tikv/pd/client/http"
 	"github.com/tikv/pd/client/pkg/utils/grpcutil"
 	"github.com/tikv/pd/client/pkg/utils/tlsutil"
-	"github.com/tikv/pd/pkg/codec"
-	"github.com/tikv/pd/pkg/mcs/utils"
+	mcsutils "github.com/tikv/pd/pkg/mcs/utils"
 	"github.com/tikv/pd/pkg/statistics"
 	"github.com/tikv/pd/pkg/utils/logutil"
 	"github.com/tikv/pd/tools/pd-heartbeat-bench/config"
 	"github.com/tikv/pd/tools/pd-heartbeat-bench/metrics"
+	"github.com/tikv/pd/tools/utils"
 )
 
 const (
-	bytesUnit            = 128
-	keysUint             = 8
-	queryUnit            = 8
-	hotByteUnit          = 16 * units.KiB
-	hotKeysUint          = 256
-	hotQueryUnit         = 256
 	regionReportInterval = 60 // 60s
 	storeReportInterval  = 10 // 10s
 	capacity             = 4 * units.TiB
@@ -183,149 +176,6 @@ func putStores(ctx context.Context, cfg *config.Config, cli pdpb.PDClient, store
 	}
 }
 
-// Regions simulates all regions to heartbeat.
-type Regions struct {
-	regions       []*pdpb.RegionHeartbeatRequest
-	awakenRegions atomic.Value
-
-	updateRound int
-
-	updateLeader []int
-	updateEpoch  []int
-	updateSpace  []int
-	updateFlow   []int
-}
-
-func (rs *Regions) init(cfg *config.Config) {
-	rs.regions = make([]*pdpb.RegionHeartbeatRequest, 0, cfg.RegionCount)
-	rs.updateRound = 0
-
-	// Generate regions
-	id := uint64(1)
-	now := uint64(time.Now().Unix())
-
-	for i := range cfg.RegionCount {
-		region := &pdpb.RegionHeartbeatRequest{
-			Header: header(),
-			Region: &metapb.Region{
-				Id:          id,
-				StartKey:    codec.GenerateTableKey(int64(i)),
-				EndKey:      codec.GenerateTableKey(int64(i + 1)),
-				RegionEpoch: &metapb.RegionEpoch{ConfVer: 2, Version: maxVersion},
-			},
-			ApproximateSize: bytesUnit,
-			Interval: &pdpb.TimeInterval{
-				StartTimestamp: now,
-				EndTimestamp:   now + regionReportInterval,
-			},
-			QueryStats:      &pdpb.QueryStats{},
-			ApproximateKeys: keysUint,
-			Term:            1,
-		}
-		id += 1
-		if i == 0 {
-			region.Region.StartKey = []byte("")
-		}
-		if i == cfg.RegionCount-1 {
-			region.Region.EndKey = []byte("")
-		}
-
-		peers := make([]*metapb.Peer, 0, cfg.Replica)
-		for j := range cfg.Replica {
-			peers = append(peers, &metapb.Peer{Id: id, StoreId: uint64((i+j)%cfg.StoreCount + 1)})
-			id += 1
-		}
-
-		region.Region.Peers = peers
-		region.Leader = peers[0]
-		rs.regions = append(rs.regions, region)
-	}
-}
-
-func (rs *Regions) update(cfg *config.Config, options *config.Options) {
-	rs.updateRound += 1
-
-	// Generate sample index
-	indexes := make([]int, cfg.RegionCount)
-	for i := range indexes {
-		indexes[i] = i
-	}
-	reportRegions := pick(indexes, cfg.RegionCount, options.GetReportRatio())
-
-	reportCount := len(reportRegions)
-	rs.updateFlow = pick(reportRegions, reportCount, options.GetFlowUpdateRatio())
-	rs.updateLeader = randomPick(reportRegions, reportCount, options.GetLeaderUpdateRatio())
-	rs.updateEpoch = randomPick(reportRegions, reportCount, options.GetEpochUpdateRatio())
-	rs.updateSpace = randomPick(reportRegions, reportCount, options.GetSpaceUpdateRatio())
-	var (
-		updatedStatisticsMap = make(map[int]*pdpb.RegionHeartbeatRequest)
-		awakenRegions        []*pdpb.RegionHeartbeatRequest
-	)
-
-	// update leader
-	for _, i := range rs.updateLeader {
-		region := rs.regions[i]
-		region.Leader = region.Region.Peers[rs.updateRound%cfg.Replica]
-	}
-	// update epoch
-	for _, i := range rs.updateEpoch {
-		region := rs.regions[i]
-		region.Region.RegionEpoch.Version += 1
-		if region.Region.RegionEpoch.Version > maxVersion {
-			maxVersion = region.Region.RegionEpoch.Version
-		}
-	}
-	// update space
-	for _, i := range rs.updateSpace {
-		region := rs.regions[i]
-		region.ApproximateSize = uint64(bytesUnit * rand.Float64())
-		region.ApproximateKeys = uint64(keysUint * rand.Float64())
-	}
-	// update flow
-	for _, i := range rs.updateFlow {
-		region := rs.regions[i]
-		if region.Leader.StoreId <= uint64(options.GetHotStoreCount()) {
-			region.BytesWritten = uint64(hotByteUnit * (1 + rand.Float64()) * 60)
-			region.BytesRead = uint64(hotByteUnit * (1 + rand.Float64()) * 10)
-			region.KeysWritten = uint64(hotKeysUint * (1 + rand.Float64()) * 60)
-			region.KeysRead = uint64(hotKeysUint * (1 + rand.Float64()) * 10)
-			region.QueryStats = &pdpb.QueryStats{
-				Get: uint64(hotQueryUnit * (1 + rand.Float64()) * 10),
-				Put: uint64(hotQueryUnit * (1 + rand.Float64()) * 60),
-			}
-		} else {
-			region.BytesWritten = uint64(bytesUnit * rand.Float64())
-			region.BytesRead = uint64(bytesUnit * rand.Float64())
-			region.KeysWritten = uint64(keysUint * rand.Float64())
-			region.KeysRead = uint64(keysUint * rand.Float64())
-			region.QueryStats = &pdpb.QueryStats{
-				Get: uint64(queryUnit * rand.Float64()),
-				Put: uint64(queryUnit * rand.Float64()),
-			}
-		}
-		updatedStatisticsMap[i] = region
-	}
-	// update interval
-	for _, region := range rs.regions {
-		region.Interval.StartTimestamp = region.Interval.EndTimestamp
-		region.Interval.EndTimestamp = region.Interval.StartTimestamp + regionReportInterval
-	}
-	for _, i := range reportRegions {
-		region := rs.regions[i]
-		// reset the statistics of the region which is not updated
-		if _, exist := updatedStatisticsMap[i]; !exist {
-			region.BytesWritten = 0
-			region.BytesRead = 0
-			region.KeysWritten = 0
-			region.KeysRead = 0
-			region.QueryStats = &pdpb.QueryStats{}
-		}
-		awakenRegions = append(awakenRegions, region)
-	}
-
-	rs.awakenRegions.Store(awakenRegions)
-}
-
 func createHeartbeatStream(ctx context.Context, cfg *config.Config) (pdpb.PDClient, pdpb.PD_RegionHeartbeatClient) {
 	cli, err := newClient(ctx, cfg)
 	if err != nil {
@@ -345,43 +195,6 @@ func createHeartbeatStream(ctx context.Context, cfg *config.Config) (pdpb.PDClie
 	return cli, stream
 }
 
-func (rs *Regions) handleRegionHeartbeat(wg *sync.WaitGroup, stream pdpb.PD_RegionHeartbeatClient, storeID uint64, rep report.Report) {
-	defer wg.Done()
-	var regions, toUpdate []*pdpb.RegionHeartbeatRequest
-	updatedRegions := rs.awakenRegions.Load()
-	if updatedRegions == nil {
-		toUpdate = rs.regions
-	} else {
-		toUpdate = updatedRegions.([]*pdpb.RegionHeartbeatRequest)
-	}
-	for _, region := range toUpdate {
-		if region.Leader.StoreId != storeID {
-			continue
-		}
-		regions = append(regions, region)
-	}
-
-	start := time.Now()
-	var err error
-	for _, region := range regions {
-		err = stream.Send(region)
-		rep.Results() <- report.Result{Start: start, End: time.Now(), Err: err}
-		if err == io.EOF {
-			log.Error("receive eof error", zap.Uint64("store-id", storeID), zap.Error(err))
-			err := stream.CloseSend()
-			if err != nil {
-				log.Error("fail to close stream", zap.Uint64("store-id", storeID), zap.Error(err))
-			}
-			return
-		}
-		if err != nil {
-			log.Error("send result error", zap.Uint64("store-id", storeID), zap.Error(err))
-			return
-		}
-	}
-	log.Info("store finish one round region heartbeat", zap.Uint64("store-id", storeID), zap.Duration("cost-time", time.Since(start)), zap.Int("reported-region-count", len(regions)))
-}
-
 // Stores contains store stats with lock.
 type Stores struct {
 	stat []atomic.Value
@@ -399,7 +212,7 @@ func (s *Stores) heartbeat(ctx context.Context, cli pdpb.PDClient, storeID uint6
 	cli.StoreHeartbeat(cctx, &pdpb.StoreHeartbeatRequest{Header: header(), Stats: s.stat[storeID].Load().(*pdpb.StoreStats)})
 }
 
-func (s *Stores) update(rs *Regions) {
+func (s *Stores) update(rs *utils.Regions) {
 	stats := make([]*pdpb.StoreStats, len(s.stat))
 	now := uint64(time.Now().Unix())
 	for i := range stats {
@@ -416,9 +229,9 @@ func (s *Stores) update(rs *Regions) {
 		}
 	}
 	var toUpdate []*pdpb.RegionHeartbeatRequest
-	updatedRegions := rs.awakenRegions.Load()
+	updatedRegions := rs.AwakenRegions.Load()
 	if updatedRegions == nil {
-		toUpdate = rs.regions
+		toUpdate = rs.Regions
 	} else {
 		toUpdate = updatedRegions.([]*pdpb.RegionHeartbeatRequest)
 	}
@@ -450,17 +263,6 @@ func (s *Stores) update(rs *Regions) {
 	for i := range stats {
 		s.stat[i].Store(stats[i])
 	}
-}
-
-func randomPick(slice []int, total int, ratio float64) []int {
-	rand.Shuffle(total, func(i, j int) {
-		slice[i], slice[j] = slice[j], slice[i]
-	})
-	return append(slice[:0:0], slice[0:int(float64(total)*ratio)]...)
-}
-
-func pick(slice []int, total int, ratio float64) []int {
-	return append(slice[:0:0], slice[0:int(float64(total)*ratio)]...)
 }
 
 func main() {
@@ -510,8 +312,7 @@ func main() {
 
 	initClusterID(ctx, cli)
 	go runHTTPServer(cfg, options)
-	regions := new(Regions)
-	regions.init(cfg)
+	regions := utils.NewRegions(cfg.RegionCount, cfg.Replica, header())
 	log.Info("finish init regions")
 	stores := newStores(cfg.StoreCount)
 	stores.update(regions)
@@ -536,7 +337,7 @@ func main() {
 	for {
 		select {
 		case <-heartbeatTicker.C:
-			if cfg.Round != 0 && regions.updateRound > cfg.Round {
+			if cfg.Round != 0 && regions.UpdateRound > cfg.Round {
 				exit(0)
 			}
 			rep := newReport(cfg)
@@ -547,22 +348,22 @@ func main() {
 			for i := 1; i <= cfg.StoreCount; i++ {
 				id := uint64(i)
 				wg.Add(1)
-				go regions.handleRegionHeartbeat(wg, streams[id], id, rep)
+				go regions.HandleRegionHeartbeat(wg, streams[id], id, rep)
 			}
 			if withMetric {
-				metrics.CollectMetrics(regions.updateRound, time.Second)
+				metrics.CollectMetrics(regions.UpdateRound, time.Second)
 			}
 			wg.Wait()
 
 			since := time.Since(startTime).Seconds()
 			close(rep.Results())
-			regions.result(cfg.RegionCount, since)
+			regions.Result(cfg.RegionCount, since)
 			stats := <-r
 			log.Info("region heartbeat stats",
 				metrics.RegionFields(stats, zap.Uint64("max-epoch-version", maxVersion))...)
 			log.Info("store heartbeat stats", zap.String("max", fmt.Sprintf("%.4fs", since)))
 			metrics.CollectRegionAndStoreStats(&stats, &since)
-			regions.update(cfg, options)
+			regions.Update(cfg.RegionCount, cfg.Replica, options)
 			go stores.update(regions) // update stores in background, unusually region heartbeat is slower than store update.
 		case <-resolvedTSTicker.C:
 			wg := &sync.WaitGroup{}
@@ -625,42 +426,13 @@ func newReport(cfg *config.Config) report.Report {
 	return report.NewReport(p)
 }
 
-func (rs *Regions) result(regionCount int, sec float64) {
-	if rs.updateRound == 0 {
-		// There was no difference in the first round
-		return
-	}
-
-	updated := make(map[int]struct{})
-	for _, i := range rs.updateLeader {
-		updated[i] = struct{}{}
-	}
-	for _, i := range rs.updateEpoch {
-		updated[i] = struct{}{}
-	}
-	for _, i := range rs.updateSpace {
-		updated[i] = struct{}{}
-	}
-	for _, i := range rs.updateFlow {
-		updated[i] = struct{}{}
-	}
-	inactiveCount := regionCount - len(updated)
-
-	log.Info("update speed of each category", zap.String("rps", fmt.Sprintf("%.4f", float64(regionCount)/sec)),
-		zap.String("save-tree", fmt.Sprintf("%.4f", float64(len(rs.updateLeader))/sec)),
-		zap.String("save-kv", fmt.Sprintf("%.4f", float64(len(rs.updateEpoch))/sec)),
-		zap.String("save-space", fmt.Sprintf("%.4f", float64(len(rs.updateSpace))/sec)),
-		zap.String("save-flow", fmt.Sprintf("%.4f", float64(len(rs.updateFlow))/sec)),
-		zap.String("skip", fmt.Sprintf("%.4f", float64(inactiveCount)/sec)))
-}
-
 func runHTTPServer(cfg *config.Config, options *config.Options) {
 	gin.SetMode(gin.ReleaseMode)
 	engine := gin.New()
 	engine.Use(gin.Recovery())
 	engine.Use(cors.Default())
 	engine.Use(gzip.Gzip(gzip.DefaultCompression))
-	engine.GET("metrics", utils.PromHandler())
+	engine.GET("metrics", mcsutils.PromHandler())
 	// profile API
 	pprof.Register(engine)
 	engine.PUT("config", func(c *gin.Context) {
