+++ tools/pd-tso-bench/main.go
@@ -18,9 +18,7 @@ import (
 	"context"
 	"flag"
 	"fmt"
-	"io"
 	"math/rand"
-	"net/http"
 	"net/http/httptest"
 	"os"
 	"os/signal"
@@ -28,7 +26,6 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/influxdata/tdigest"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"go.uber.org/zap"
 	"google.golang.org/grpc"
@@ -40,6 +37,7 @@ import (
 	pd "github.com/tikv/pd/client"
 	"github.com/tikv/pd/client/opt"
 	"github.com/tikv/pd/client/pkg/caller"
+	"github.com/tikv/pd/tools/utils"
 )
 
 const (
@@ -70,14 +68,6 @@ var (
 )
 
 var promServer *httptest.Server
-
-func collectMetrics(server *httptest.Server) string {
-	time.Sleep(1100 * time.Millisecond)
-	res, _ := http.Get(server.URL)
-	body, _ := io.ReadAll(res.Body)
-	res.Body.Close()
-	return string(body)
-}
 
 func main() {
 	flag.Parse()
@@ -141,7 +131,7 @@ func bench(mainCtx context.Context) {
 	}
 
 	wg.Add(1)
-	go showStats(ctx, durCh)
+	go utils.ShowStats(ctx, &wg, durCh, *interval, *verbose, promServer)
 
 	timer := time.NewTimer(*duration)
 	defer timer.Stop()
@@ -157,198 +147,6 @@ func bench(mainCtx context.Context) {
 	for _, pdCli := range pdClients {
 		pdCli.Close()
 	}
-}
-
-var latencyTDigest *tdigest.TDigest = tdigest.New()
-
-func showStats(ctx context.Context, durCh chan time.Duration) {
-	defer wg.Done()
-
-	statCtx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	ticker := time.NewTicker(*interval)
-	defer ticker.Stop()
-
-	s := newStats()
-	total := newStats()
-
-	fmt.Println()
-	for {
-		select {
-		case <-ticker.C:
-			// runtime.GC()
-			if *verbose {
-				fmt.Println(s.counter())
-			}
-			total.merge(s)
-			s = newStats()
-		case d := <-durCh:
-			s.update(d)
-		case <-statCtx.Done():
-			fmt.Println("\nTotal:")
-			fmt.Println(total.counter())
-			fmt.Println(total.percentage())
-			// Calculate the percentiles by using the tDigest algorithm.
-			fmt.Printf("P0.5: %.4fms, P0.8: %.4fms, P0.9: %.4fms, P0.99: %.4fms\n\n", latencyTDigest.Quantile(0.5), latencyTDigest.Quantile(0.8), latencyTDigest.Quantile(0.9), latencyTDigest.Quantile(0.99))
-			if *verbose {
-				fmt.Println(collectMetrics(promServer))
-			}
-			return
-		}
-	}
-}
-
-const (
-	twoDur          = time.Millisecond * 2
-	fiveDur         = time.Millisecond * 5
-	tenDur          = time.Millisecond * 10
-	thirtyDur       = time.Millisecond * 30
-	fiftyDur        = time.Millisecond * 50
-	oneHundredDur   = time.Millisecond * 100
-	twoHundredDur   = time.Millisecond * 200
-	fourHundredDur  = time.Millisecond * 400
-	eightHundredDur = time.Millisecond * 800
-	oneThousandDur  = time.Millisecond * 1000
-)
-
-type stats struct {
-	maxDur          time.Duration
-	minDur          time.Duration
-	totalDur        time.Duration
-	count           int
-	submilliCnt     int
-	milliCnt        int
-	twoMilliCnt     int
-	fiveMilliCnt    int
-	tenMSCnt        int
-	thirtyCnt       int
-	fiftyCnt        int
-	oneHundredCnt   int
-	twoHundredCnt   int
-	fourHundredCnt  int
-	eightHundredCnt int
-	oneThousandCnt  int
-}
-
-func newStats() *stats {
-	return &stats{
-		minDur: time.Hour,
-		maxDur: 0,
-	}
-}
-
-func (s *stats) update(dur time.Duration) {
-	s.count++
-	s.totalDur += dur
-	latencyTDigest.Add(float64(dur.Nanoseconds())/1e6, 1)
-
-	if dur > s.maxDur {
-		s.maxDur = dur
-	}
-	if dur < s.minDur {
-		s.minDur = dur
-	}
-
-	if dur > oneThousandDur {
-		s.oneThousandCnt++
-		return
-	}
-
-	if dur > eightHundredDur {
-		s.eightHundredCnt++
-		return
-	}
-
-	if dur > fourHundredDur {
-		s.fourHundredCnt++
-		return
-	}
-
-	if dur > twoHundredDur {
-		s.twoHundredCnt++
-		return
-	}
-
-	if dur > oneHundredDur {
-		s.oneHundredCnt++
-		return
-	}
-
-	if dur > fiftyDur {
-		s.fiftyCnt++
-		return
-	}
-
-	if dur > thirtyDur {
-		s.thirtyCnt++
-		return
-	}
-
-	if dur > tenDur {
-		s.tenMSCnt++
-		return
-	}
-
-	if dur > fiveDur {
-		s.fiveMilliCnt++
-		return
-	}
-
-	if dur > twoDur {
-		s.twoMilliCnt++
-		return
-	}
-
-	if dur > time.Millisecond {
-		s.milliCnt++
-		return
-	}
-
-	s.submilliCnt++
-}
-
-func (s *stats) merge(other *stats) {
-	if s.maxDur < other.maxDur {
-		s.maxDur = other.maxDur
-	}
-	if s.minDur > other.minDur {
-		s.minDur = other.minDur
-	}
-
-	s.count += other.count
-	s.totalDur += other.totalDur
-	s.submilliCnt += other.submilliCnt
-	s.milliCnt += other.milliCnt
-	s.twoMilliCnt += other.twoMilliCnt
-	s.fiveMilliCnt += other.fiveMilliCnt
-	s.tenMSCnt += other.tenMSCnt
-	s.thirtyCnt += other.thirtyCnt
-	s.fiftyCnt += other.fiftyCnt
-	s.oneHundredCnt += other.oneHundredCnt
-	s.twoHundredCnt += other.twoHundredCnt
-	s.fourHundredCnt += other.fourHundredCnt
-	s.eightHundredCnt += other.eightHundredCnt
-	s.oneThousandCnt += other.oneThousandCnt
-}
-
-func (s *stats) counter() string {
-	return fmt.Sprintf(
-		"count: %d, max: %.4fms, min: %.4fms, avg: %.4fms\n<1ms: %d, >1ms: %d, >2ms: %d, >5ms: %d, >10ms: %d, >30ms: %d, >50ms: %d, >100ms: %d, >200ms: %d, >400ms: %d, >800ms: %d, >1s: %d",
-		s.count, float64(s.maxDur.Nanoseconds())/float64(time.Millisecond), float64(s.minDur.Nanoseconds())/float64(time.Millisecond), float64(s.totalDur.Nanoseconds())/float64(s.count)/float64(time.Millisecond),
-		s.submilliCnt, s.milliCnt, s.twoMilliCnt, s.fiveMilliCnt, s.tenMSCnt, s.thirtyCnt, s.fiftyCnt, s.oneHundredCnt, s.twoHundredCnt, s.fourHundredCnt,
-		s.eightHundredCnt, s.oneThousandCnt)
-}
-
-func (s *stats) percentage() string {
-	return fmt.Sprintf(
-		"count: %d, <1ms: %2.2f%%, >1ms: %2.2f%%, >2ms: %2.2f%%, >5ms: %2.2f%%, >10ms: %2.2f%%, >30ms: %2.2f%%, >50ms: %2.2f%%, >100ms: %2.2f%%, >200ms: %2.2f%%, >400ms: %2.2f%%, >800ms: %2.2f%%, >1s: %2.2f%%", s.count,
-		s.calculate(s.submilliCnt), s.calculate(s.milliCnt), s.calculate(s.twoMilliCnt), s.calculate(s.fiveMilliCnt), s.calculate(s.tenMSCnt), s.calculate(s.thirtyCnt), s.calculate(s.fiftyCnt),
-		s.calculate(s.oneHundredCnt), s.calculate(s.twoHundredCnt), s.calculate(s.fourHundredCnt), s.calculate(s.eightHundredCnt), s.calculate(s.oneThousandCnt))
-}
-
-func (s *stats) calculate(count int) float64 {
-	return float64(count) * 100 / float64(s.count)
 }
 
 func reqWorker(ctx context.Context, pdClients []pd.Client, clientIdx int, durCh chan time.Duration) {
